Excecoes

+++ src/serial.h
@@ -14,7 +14,7 @@
 
 extern int serial_init(char p[], uint32_t b);
 extern int serial_close(void);
-extern size_t serial_transaction(uint8_t *tx, uint8_t *rx, uint16_t msg_size,
+extern int serial_transaction(uint8_t *tx, uint8_t *rx, uint16_t msg_size,
 		uint16_t resp_size);
 
 #endif /* SERIAL_H_ */
